Reporte/AZURE/1extract/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
PyMuPDF
azure-storage-blob
pandas
pyarrow